safari-534.53-branch/LayoutTests/svg/animations/animate-color-calcMode-discrete-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Test calcMode discrete with from-to animation on colors. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
